+++ src/main/index.ts
@@ -5,6 +5,7 @@ import icon from '../../resources/icon.png?asset'
 import { shardManager } from '@shared/yuekui-shard/manager'
 import './shards/Hello-world-shard/index'
 import './shards/Lcu-connect/index'
+import './shards/Auto-pick-shards/index'
 
 function createWindow(): void {
   // Create the browser window.